requirements.txt
mutagen
pillow
soundfile
sounddevice
pythonnet
numpy
scipy
https://github.com/kivymd/KivyMD/archive/master.zip
pyside6
